requirements.txt
anyascii==0.3.0
APScheduler==3.8.1
asgiref==3.4.1
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.10
click==8.0.3
colorhash==1.0.4
configparser==5.2.0
cryptography==3.4.7
Django==3.2.11
django-filter==21.1
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fabric==2.6.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-MonitoringDashboard==3.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
invoke==1.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
l18n==2021.3
MarkupSafe==2.0.1
numpy==1.22.1
openpyxl==3.0.9
paramiko==2.7.2
pathlib2==2.3.5
Pillow==8.4.0
psutil==5.9.0
pycodestyle==2.8.0
pycparser==2.20
PyNaCl==1.4.0
python-bcrypt==0.3.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
toml==0.10.2
tqdm==4.61.2
tzlocal==2.0.0
urllib3==1.26.8
vsco-scraper==0.53
wagtail==2.15.1
wagtail-nhsuk-frontend==1.2.2
webencodings==0.5.1
Werkzeug==2.0.2
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
